Add 'generate_epilog' function to provide argparse with usage examples
This commit adds a new function in __main__.py named 'generate_epilog'. This helper function produces usage examples that are populated within the argparse class property 'epilog'. This change also includes adding 'argparse.RawDescriptionHelpFormatter' as a value for the argparse clss property 'formatter_class' to ensure the output is formatted properly when it's printing to the terminal.